adding profile policies to UserPolicy

+++ app/Http/Controllers/User/ProfileController.php
@@ -4,12 +4,18 @@ namespace App\Http\Controllers\User;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\User\Profile\UpdateProfileRequest;
+use App\Models\User;
+use App\Policies\ProfilePolicy;
 use function view;
 
 class ProfileController extends Controller
 {
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function show()
     {
+        $this->authorize('viewProfile', User::class);
         return view('dashboard.users.profile.show', [
             'user' => \Auth::user()
         ]);
@@ -17,6 +23,7 @@ class ProfileController extends Controller
 
     public function edit()
     {
+        $this->authorize('updateProfile', User::class);
         return view('dashboard.users.profile.edit', [
             'user' => \Auth::user()
         ]);
@@ -24,6 +31,7 @@ class ProfileController extends Controller
 
     public function update(UpdateProfileRequest $request, $id)
     {
+        $this->authorize('updateProfile', User::class);
         try {
             \Auth::user()->update($request->validated());
         } catch (\Exception $exception) {
@@ -35,6 +43,7 @@ class ProfileController extends Controller
 
     public function destroy()
     {
+        $this->authorize('deleteAccount', User::class);
         try {
             \Auth::user()->delete();
         } catch (\Exception $exception) {
